Split WidgetHostingService and WidgetServiceService (#2610)
