+++ tools/Dashboard/DevHome.Dashboard/Extensions/ServiceExtensions.cs
@@ -24,6 +24,7 @@ public static class ServiceExtensions
                 ActivatorUtilities.CreateInstance<WidgetViewModel>(sp, widget, widgetSize, widgetDefinition));
 
         // Services
+        services.AddSingleton<IWidgetServiceService, WidgetServiceService>();
         services.AddSingleton<IWidgetHostingService, WidgetHostingService>();
         services.AddSingleton<IWidgetIconService, WidgetIconService>();
         services.AddSingleton<IWidgetScreenshotService, WidgetScreenshotService>();
