+++ tools/Dashboard/DevHome.Dashboard/Services/IWidgetServiceService.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Threading.Tasks;
+using static DevHome.Dashboard.Services.WidgetServiceService;
+
+namespace DevHome.Dashboard.Services;
+
+public interface IWidgetServiceService
+{
+    public bool CheckForWidgetServiceAsync();
+
+    public Task<bool> TryInstallingWidgetService();
+
+    public WidgetServiceStates GetWidgetServiceState();
+}
